requirements.txt
requests
pysoundfile
numpy
pandas
geopandas
flask